requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.8.0
ase==3.22.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
chemparse==0.3.1
chex==0.1.86
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.8
dm-haiku==0.0.11
dnspython==2.6.1
element-coder==0.0.8
emmet-core==0.68.0
etils==1.7.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.13.1
flax==0.8.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
hydra-core==1.3.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
jmp==0.0.4
joblib==1.3.2
kiwisolver==1.4.5
latexcodec==3.0.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matbench-genmetrics==0.6.1
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
monty==2024.10.21
mp-api==0.36.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.1.8
orbax-checkpoint==0.5.7
p-tqdm==1.4.0
packaging==24.0
palettable==3.3.3
pandas==2.2.1
pathos==0.3.2
pillow==10.2.0
plotly==5.20.0
pox==0.3.4
ppft==1.7.6.8
protobuf==5.26.0
psutil==6.0.0
pybtex==0.24.0
pydantic==1.10.18
Pygments==2.17.2
pymatgen==2024.10.3
pymatviz==0.8.1
pymongo==4.6.2
pyparsing==3.1.2
pystow==0.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxtal==0.6.3
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.14.1
six==1.16.0
SMACT==2.5.5
sniffio==1.3.1
spglib==2.5.0
starlette==0.45.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorstore==0.1.56
threadpoolctl==3.4.0
toolz==0.12.1
tqdm==4.66.2
triton==2.3.0
typing_extensions==4.10.0
tzdata==2024.1
uncertainties==3.1.7
urllib3==2.2.1
uvicorn==0.34.0
yarl==1.9.4
zipp==3.18.1
et-xmlfile==2.0.0 
openpyxl==3.1.5
